Converted to a maven project
