+++ src/main/java/com/dkabot/DkabotShop/ManagedFile.java
